watcher_service/requirements.txt
# Watcher service dependencies
# torch / torchvision / torchaudio are installed separately in Dockerfile
# (requires --index-url https://download.pytorch.org/whl/cu128)

# ── File watching ─────────────────────────────────────────────────────────────
watchdog>=4.0.0
python-dotenv>=1.0.0
pyyaml>=6.0

# ── VideoSeal core deps (no strict pin — let pip resolve for Python 3.10) ────
av>=14.0.0
omegaconf>=2.3.0
einops>=0.6.0
timm==0.9.16
safetensors>=0.4.0
transformers>=4.30.0
lpips>=0.1.4
PyWavelets>=1.3.0
scikit-image>=0.21.0
scipy>=1.11.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# ── ECC ───────────────────────────────────────────────────────────────────────
reedsolo>=1.7.0
bchlib==2.1.3

# ── AudioSeal deps ────────────────────────────────────────────────────────────
huggingface-hub>=0.20.0

# ── Extract App (web UI) ─────────────────────────────────────────────────────
flask>=3.0.0

# ── Misc ──────────────────────────────────────────────────────────────────────
requests>=2.28.0
tqdm>=4.60.0
pandas>=2.0.0
psutil>=5.9.0
